requirements.txt
chainlit==1.1.305
pypdf==4.2.0
langchain==0.2.6
langchain-community==0.2.6
openai==1.35.5
python-dotenv==1.0.1
pinecone-client==4.1.1
tiktoken==0.3.3
langchain-openai==0.1.10